% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{designed_prems}
\alias{designed_prems}
\title{Filter the designed premises}
\usage{
designed_prems(
  g,
  cables,
  w = sf::st_sf(geometry = sf::st_sfc(sf::st_multipoint()))
)
}
\arguments{
\item{g}{= grey premises (sf)}

\item{w}{= white premises (sf)}

\item{cables=cables}{(sf)}
}
\value{
sf of designed premises (all whites + connected greys)
}
\description{
Filter the designed premises
}
