% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{net_arcgisread}
\alias{net_arcgisread}
\title{Read a layer from an Arcgis tool input}
\usage{
net_arcgisread(x, precision = 100)
}
\arguments{
\item{x}{the list of layers in the param_in of the toolbox}

\item{precision}{The precision as per sf objects. Default 100 (greater values get better precision)}
}
\value{
list of the tool input layers as sf objects
}
\description{
Read a layer from an Arcgis tool input
}
\examples{
net_arcgisread(x)
}
