% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{net_arcread}
\alias{net_arcread}
\title{Read an online hosted feature layer}
\usage{
net_arcread(path, precision = 100)
}
\arguments{
\item{path}{The path of the online hosted feature layer}

\item{precision}{The precision as per sf objects. Default 100 (greater values get better precision)}
}
\value{
sflayer The layer in sf format
}
\description{
Read an online hosted feature layer
}
\examples{
net_arcread("https://services8.arcgis.com/8nY6VQZZM2Z9cUt0/arcgis/rest/services/HALO_BLG_1_Feeder_Cable_Route/FeatureServer")
}
