% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{net_boundaries}
\alias{net_boundaries}
\title{Create Network graph from boundaries as inputs (the parent of a boundary is the smallest container)}
\usage{
net_boundaries(b, directed = TRUE)
}
\arguments{
\item{b}{= boundaries layer in sf format}

\item{directed}{TRUE if you want a directed graph(default)}
}
\value{
graph Network as tidygraph graph of boundaries (as nodes)
}
\description{
Create Network graph from boundaries as inputs (the parent of a boundary is the smallest container)
}
