% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{net_bplot}
\alias{net_bplot}
\title{Create simple Network plot for boundaries}
\usage{
net_bplot(
  x,
  name = "temp.svg",
  directory = paste0(getwd(), "/boundaries"),
  layout = "tree"
)
}
\arguments{
\item{x}{graph of tidygraph class (the nodes must contain the label_text field with the string for plot)}

\item{name}{the name of the file output (the output is an svg file, so a name ending as '.svg' is better)}

\item{directory}{the place where to place the plot}

\item{layout}{type of the graph}
}
\value{
graph Network preliminary plot
}
\description{
Create simple Network plot for boundaries
}
