% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{net_cables}
\alias{net_cables}
\title{Create Network graph from cables as inputs (only network cables on the graph)}
\usage{
net_cables(x, directed = TRUE)
}
\arguments{
\item{x}{= Cables layer in sf format}

\item{directed}{TRUE if you want a directed graph(default)}
}
\value{
graph Network as tidygraph graph of cables (as edges)
}
\description{
Create Network graph from cables as inputs (only network cables on the graph)
}
