% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{net_closures}
\alias{net_closures}
\title{Create Network graph from closures sf layer as input}
\usage{
net_closures(x, directed = TRUE)
}
\arguments{
\item{x}{= closures layer in sf format}

\item{directed}{TRUE if you want a directed graph(default)}
}
\value{
a graph of closure layer (no edges)
}
\description{
Create Network graph from closures sf layer as input
}
