% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{net_devices}
\alias{net_devices}
\title{Take the splice closure sf and mutate it to include the capacity}
\usage{
net_devices(x)
}
\arguments{
\item{x}{= closures layer in sf format}
}
\value{
sf of devices(pivot longer of splice closures)
}
\description{
Take the splice closure sf and mutate it to include the capacity
}
