% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{net_edges2tibble}
\alias{net_edges2tibble}
\title{Convert net edges to tibble}
\usage{
net_edges2tibble(net)
}
\arguments{
\item{net}{Network graph}
}
\value{
tibble rapresentation of the edges
}
\description{
Convert net edges to tibble
}
