% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{net_filter}
\alias{net_filter}
\title{Filter the net (wrapper around tidygraph to_subgraph())}
\usage{
net_filter(net, ...)
}
\arguments{
\item{net}{Network graph}

\item{...}{Parameter to be passed to filter function}

\item{premises}{Premises layer in sf format}
}
\value{
graph subNetwork as tidygraph graph
}
\description{
Filter the net (wrapper around tidygraph to_subgraph())
}
