% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{net_graph}
\alias{net_graph}
\title{Create Network graph from layers as inputs}
\usage{
net_graph(cables, closures, directed = TRUE, cables_as_nodes = FALSE)
}
\arguments{
\item{cables}{Cables layer in sf format}

\item{closures}{Closures layer in sf format}

\item{directed}{TRUE if you want a directed graph(default)}

\item{cables_as_nodes}{if TRUE the cables are represented as nodes along with the devices, if FALSE(default) the cables are represented as edges}
}
\value{
graph Network as tidygraph graph (the nodes table has a group column for connected subgraphs)
}
\description{
Create Network graph from layers as inputs
}
