% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{net_graph2}
\alias{net_graph2}
\title{Create complete Network graph}
\usage{
net_graph2(cables, closures, premises, boundaries)
}
\arguments{
\item{cables}{sf layer}

\item{closures}{sf layer}

\item{premises}{sf layer}

\item{boundaries}{sf layer}
}
\value{
graph Network as tidygraph object
}
\description{
Create complete Network graph
}
