% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{net_load}
\alias{net_load}
\title{Add load to each boundary of a net_boundaries}
\usage{
net_load(bnet, dp)
}
\arguments{
\item{bnet}{= boundaries net (graph)}

\item{dp}{= designed premises}

\item{cl}{= splice closures}
}
\value{
boundaries net with load on nodes
}
\description{
Add load to each boundary of a net_boundaries
}
