% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{net_multinodes}
\alias{net_multinodes}
\title{Search for nodes that shares the same geometries}
\usage{
net_multinodes(net)
}
\arguments{
\item{net}{Network graph}
}
\value{
list of multiple nodes indexes
}
\description{
Search for nodes that shares the same geometries
}
