% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{net_nodes2sf}
\alias{net_nodes2sf}
\title{Convert net nodes to sf}
\usage{
net_nodes2sf(net)
}
\arguments{
\item{net}{Network graph}
}
\value{
sf rapresentation of the nodes
}
\description{
Convert net nodes to sf
}
