% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{net_nodes2tibble}
\alias{net_nodes2tibble}
\title{Convert net nodes to tibble}
\usage{
net_nodes2tibble(net)
}
\arguments{
\item{net}{Network graph}
}
\value{
tibble rapresentation of the nodes
}
\description{
Convert net nodes to tibble
}
