% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{net_plot}
\alias{net_plot}
\title{Create simple Network plot}
\usage{
net_plot(net, ...)
}
\arguments{
\item{net}{graph of tidygraph class}

\item{closures}{layout as from ggraph library}
}
\value{
graph Network preliminary plot
}
\description{
Create simple Network plot
}
