% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{net_uprn}
\alias{net_uprn}
\title{Create Network graph from premises sf layer as input}
\usage{
net_uprn(x, directed = TRUE)
}
\arguments{
\item{x}{= Premises layer in sf format}

\item{directed}{TRUE if you want a directed graph(default)}
}
\value{
a graph of premises layer (no edges)
}
\description{
Create Network graph from premises sf layer as input
}
