% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{net_virtual_cables}
\alias{net_virtual_cables}
\title{Create Network graph of virtual cables from splice closures as inputs}
\usage{
net_virtual_cables(x, directed = TRUE)
}
\arguments{
\item{x}{= splice closures in sf format}

\item{directed}{TRUE if you want a directed graph(default)}
}
\value{
graph Network of only virtual cables (as edges)
}
\description{
Create Network graph of virtual cables from splice closures as inputs
}
