% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{sld}
\alias{sld}
\title{Create SLD schema}
\usage{
sld(
  cables,
  premises,
  closures,
  boundaries,
  out_param,
  logo =
    "C:/Users/DomenicoGuadalupi/OneDrive - Viberoptix/Documents/R/SLD/viberoptix_logo_testo.png"
)
}
\arguments{
\item{cables}{Cables layer in sf format}

\item{premises}{Premises layer in sf format}

\item{closures}{Closures layer in sf format}

\item{boundaries}{Boundaries layer in sf format}

\item{out_param}{Short prefix for outputs}

\item{logo}{PNG for output logo}
}
\value{
graph Network as tidygraph graph
}
\description{
Create SLD schema
}
