% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldtool.R
\name{valid_boundaries}
\alias{valid_boundaries}
\title{Boundary validation for graph generation. The function filter the dev boundaries and add R and dev_type values. Return a subset of validated boundaries}
\usage{
valid_boundaries(b)
}
\arguments{
\item{b}{= boundaries sf}
}
\value{
subset of device valid boundaries (sf)
}
\description{
Boundary validation for graph generation. The function filter the dev boundaries and add R and dev_type values. Return a subset of validated boundaries
}
